// src/app/precisa/ventas/Reserva/model/DtoReserva.ts
export class DtoReserva {
    constructor() {
    }   
    DocumentoVen:           string;
    num:                    number;
    CompaniaCodigo:         string;
    NomProyecto:            string;
    DescripcionCorta:       string;
    IdSede:                 number;
    SedDescripcion:         string;
    DesCondicion:           string;
    DesSituacion:           string;
    Situacion:              number;
    DesUbicacion:           string;
    NomManzana:             string;
    TipoDocumento:          string;
    Documento:              string;
    Cliente:                string;
    Vendedor:               string;
    IdReserva:              number|string;
    IdManzana:              number;
    IdProyecto:             number;
    IdLote:                 number;
    IdCliente:              number;
    IdVendedor:             number;
    IdConyuge:              number;
    FlatReserva:            number;
    Codigo:                 string;
    Descripcion:            string;
    FechaReserva:           Date;
    MonedaCodigo:           string;
    DesMoneda:              string;
    Observacion:            string;
    CostoTotal:             number;
    Diametro:               number;
    CuotaInicial:           number;
    ValorSeparacion:        number = 0;
    Valor:                  number;
    FechaInicial:           Date;
    FechaFinal:             Date;
    DiasGracia:             number;
    ComisionTotal:          number;
    TipoPago:               string;
    Banco:                  string;
    Estado:                 number;
    UsuarioCreacion:        string;
    UsuarioModificacion:    string;
    FechaCreacion:          Date;
    FechaModificacion:      Date;
    IpCreacion:             string;
    IpModificacion:         string;
    DesEstado:              string;
    NomBanco:               string;
    Lote:                   string;
    CuentaBancaria:         number;
    NomCuentaBancaria:      string;
    NomTipoPago:            string;

}